+++ CHANGELOG.rst
@@ -16,7 +16,7 @@ Unreleased
 * **PyBullet integration**: added support for PyBullet client and forward/inverse kinematic solver
 * Added ``ClientInterface``, ``PlannerInterface`` and various backend feature interfaces
 * Added implementations of these interfaces for ROS and V-REP
-* Added ``attributes`` dictionary to ``Robot`` class 
+* Added ``attributes`` dictionary to ``Robot`` class
 
 **Changed**
 
@@ -27,7 +27,8 @@ Unreleased
 * Parameter ``backend`` of forward kinematics has been renamed to ``solver``
 * The signatures of all kinematics, motion planning and planning scene management methods have been homogenized across backend clients and within ``Robot``
 * All examples have been updated to reflect these changes
-
+* Renamed all ``RobotArtist`` implementations to ``RobotModelArtist`` to reflect
+  the fact they depend on ``compas.robots.RobotModel``.
 
 **Fixed**
 
@@ -38,6 +39,7 @@ Unreleased
 * The methods ``forward_kinematics``, ``inverse_kinematics``, ``plan_cartesian_motion`` and ``plan_motion``
   of ``Robot`` class have been refactored, but a backwards-compatible deprecated version with the old
   signatures still exists suffixed by ``_deprecated``, e.g. ``forward_kinematics_deprecated``.
+* ``RobotArtist`` are deprecated in favor of ``RobotModelArtist``.
 
 **Removed**